+++ CHANGELOG.md
@@ -6,9 +6,15 @@ Currently the versioning policy of this project follows [Semantic Versioning v2.
 
 ## Unreleased - 2018-??-??
 
+### Added
+
+* Support for errorprone @CheckReturnValue annotation ([#592](https://github.com/spotbugs/spotbugs/issues/592))
+
 ### Fixed
 
+* Handle annotation on `package-info.class` properly ([#592](https://github.com/spotbugs/spotbugs/issues/592))
 * Update asm to 6.1.1 to support Java 10
+* Update Apache BCEL to 6.2 to support Java 9 package & module reference
 * Close source file after analysis ([#591](https://github.com/spotbugs/spotbugs/issues/591))
 
 ## 3.1.2 - 2018-02-24